refactor: Remove app/utils.ts and move utilities to app/utils
