+++ app/components/hooks/useTotalStakedEth.ts
@@ -1,5 +1,5 @@
 import { UseQueryResult, useQuery } from "@tanstack/react-query";
-import { SupportedNetwork, TokenRecord } from "@/app/utils";
+import { SupportedNetwork, TokenRecord } from "@/app/utils/types";
 import { useShareRates } from "./useShareRates";
 import { useTotalStakedTokens } from "./useTotalStakedTokens";
