+++ app/utils/api/deposits.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosResponse } from "axios";
-import { SupportedNetwork, TokenRecord } from "@/app/utils";
+import { SupportedNetwork, TokenRecord } from "@/app/utils/types";
 
 export interface ApiDepositsEntry {
   totalAmount: number;
